weather/requirements.txt
requests
django-decouple